LayoutTests/fast/regions/region-element-display-change-expected.txt

Test that RenderObject is recreated correctly after changing display type.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testElement(element, "none") is false
PASS testElement(element, "block") is true
PASS testElement(element, "inline-block") is true
PASS testElement(element, "compact") is false
PASS testElement(element, "inline") is false
PASS testElement(element, "table") is false
PASS testElement(element, "inline-table") is false
PASS testElement(element, "table-cell") is true
PASS testElement(element, "table-caption") is true
PASS testElement(element, "list-item") is true
PASS testElement(element, "-webkit-box") is false
PASS testElement(element, "-webkit-inline-box") is false
PASS testElement(element, "-webkit-flex") is false
PASS testElement(element, "-webkit-inline-flex") is false
PASS successfullyParsed is true

TEST COMPLETE
